refactor: adjust `SerieForm` component to use in both edit and crete page
